Hotel booking site in java
Any customers need to book hotel, ins or apartment they need to visit their office which is situated in Goa for checking the availability as well as negotiation, But with the help of above sites we can book rooms, ins and food from home. It will help them to manage the customers booking easily and also to keep the customers data more safe.It will also help staffs to keep in track their customer’s online booking request as well as easily to reply feedback to the customers. For that they hired a developer to build the web system for this reputed company.